% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Landslides Warnings Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shiny::runApp()}.}
}
\value{
The Shiny app is launched; no R object is returned.
}
\description{
This function launches the Shiny application included in the
\strong{ditwahLandslide} package. The app must be located in
\verb{inst/app/} within the package source. When installed, this
directory is available via \code{system.file()}.
}
\examples{
\dontrun{
  run_app()
}

}
