% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landslide_heatplot.R
\name{viz_landslide_warnings_from_to}
\alias{viz_landslide_warnings_from_to}
\title{Heatmap of Transitions Between Landslide Warning Levels}
\usage{
viz_landslide_warnings_from_to(
  district,
  province,
  interactivity = TRUE,
  data = ditwah_landslides_warnings,
  ang = 90
)
}
\arguments{
\item{district}{District name}

\item{province}{Province name}

\item{interactivity}{if TRUE interactive plot will be generated, if FALSE static plot will be generated}

\item{data}{name of the dataset, by default ditwah_landslides_warnings}

\item{ang}{angle of the x-axis lables}
}
\value{
Heatmap plot of landslide warnings
}
\description{
Temporal progression of landslide warning state
}
\author{
Thiyanga S. Talagala, Geethma R. Jayasinghe, Y. M. Amali P. Rajapaksha
}
