% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landslide_heatplot.R
\name{viz_landslide_warnings_from_to}
\alias{viz_landslide_warnings_from_to}
\title{Heatmap of Transitions Between Landslide Warning Levels}
\usage{
viz_landslide_warnings_from_to(
  data,
  district = NULL,
  province = NULL,
  interactivity = TRUE,
  ang = 90
)
}
\arguments{
\item{data}{name of the dataset, by default ditwah_landslides_warnings}

\item{district}{District name or All, if no district put NULL}

\item{province}{Province name or All, if no province put NULL}

\item{interactivity}{if TRUE interactive plot will be generated, if FALSE static plot will be generated}

\item{ang}{angle of the x-axis lables}
}
\value{
Heatmap plot of landslide warnings
}
\description{
Temporal progression of landslide warning state
}
\author{
Thiyanga S. Talagala, Geethma R. Jayasinghe, Y. M. Amali P. Rajapaksha
}
